1. nondeterminism bug fixed: -- removed previous solution with assert_token_seen in learn_code.c -- The issue was: we can get a negative feedback with input i1, but the same word can be positive with an input i2. So far we took care of this case only if we first saw i2 and then i1 (by filtering then out; see in learn_code.c #include "positive_queries_filter.c")
but did not take care of the other side. It is not symmetric, because when you see poisitve feedback you know that's the correct answer. When you see negative you do not know, and cannot undo L* later on when you realize i2 exists. 

So it is now solved by rechecking the negative feedback, i.e., when we get a negative feedback we recast it as a query (well, actually this is done by libalf anyway but so far we suppressed it to save a query check, but not anymore). 
2. queries containing assert letter in the middle are now filtered out in the pre_check. The semantic of the assert letter is: "assert failure", hence it can only be in the end.